fix some tests, move config
