+++ http-client/src/test/groovy/io/micronaut/http/client/HostHeaderSpec.groovy
@@ -112,7 +112,8 @@ class HostHeaderSpec extends Specification {
                 'spec.name': 'HostHeaderSpec',
                 'micronaut.ssl.enabled': true,
                 'micronaut.ssl.buildSelfSigned': true,
-                'micronaut.ssl.port': 443
+                'micronaut.ssl.port': 443,
+                'micronaut.http.client.ssl.insecureTrustAllCertificates': true,
         ]).run(EmbeddedServer)
         def asyncClient = HttpClient.create(embeddedServer.getURL())
         BlockingHttpClient client = asyncClient.toBlocking()
@@ -137,9 +138,10 @@ class HostHeaderSpec extends Specification {
                 'spec.name': 'HostHeaderSpec',
                 'micronaut.ssl.port': -1,
                 'micronaut.ssl.enabled': true,
-                'micronaut.ssl.buildSelfSigned': true
+                'micronaut.ssl.buildSelfSigned': true,
+                'micronaut.http.client.ssl.insecureTrustAllCertificates': true,
         ]).run(EmbeddedServer)
-        def asyncClient = HttpClient.create(embeddedServer.getURL())
+        def asyncClient = embeddedServer.applicationContext.createBean(HttpClient, embeddedServer.getURL())
         BlockingHttpClient client = asyncClient.toBlocking()
 
         when:
