+++ http-client/src/test/groovy/io/micronaut/http/client/SslRefreshSpec.groovy
@@ -42,7 +42,8 @@ class SslRefreshSpec extends Specification {
             'micronaut.server.ssl.ciphers': ciphers,
             'micronaut.http.client.ssl.client-authentication': 'NEED',
             'micronaut.http.client.ssl.key-store.path': 'classpath:certs/client1.p12',
-            'micronaut.http.client.ssl.key-store.password': 'secret'
+            'micronaut.http.client.ssl.key-store.password': 'secret',
+            'micronaut.http.client.ssl.insecureTrustAllCertificates': true
     ]
     @Shared @AutoCleanup EmbeddedServer embeddedServer = ApplicationContext
                                                             .builder()
