+++ http-server-netty/src/test/groovy/io/micronaut/http/server/netty/http2/Http2PostTest.java
@@ -120,7 +120,8 @@ public class Http2PostTest implements TestPropertyProvider {
     public Map<String, String> getProperties() {
         return CollectionUtils.mapOf(
                 "micronaut.ssl.enabled", true,
-                "micronaut.ssl.buildSelfSigned", true
+                "micronaut.ssl.buildSelfSigned", true,
+                "micronaut.http.client.ssl.insecureTrustAllCertificates", true
         );
     }
